build: Link against libtpm2tctisock.
This removes the immediate dependency on the socket TCTI sources
from the TPM2.0-TSS source tree resolving #6.

The code in the TPM2.0-TSS that this repo builds manually still
references this code (since it's local to that source tree) and
so these headers must remain in the C*FLAGS variable via -I. We
can't resolve this yet without sorting out an approach to #5.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>