build: Link against the libtpm2sapi library.
This commit resolves #7. The include statements in the code have been
fixed up to use the installed headers. The bootstrap script has been
to no longer sets up the build in the TSS source repo. The build target
for libtpm.a has been removed as well so no source in the TSS source
tree is built. Finally the INSTALL file has been updated to remove the
dependency on autoconf archive and add a note on the new dependency on
the SAPI library and headers.

We still some very brittle dependencies on the sources in the
TPM2.0-TSS source tree but at least now we don't have to do a full
build of the static TSS library. We do however still build a number of
the convenience / common source files from the tss tree and link those
in.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>