build: Build common sources into library.
We were building every object into every executable which is pretty
inefficient. By building these common files into a library we the
linker will selectively pull in the bits we need. You can see this
just in the decreased size of the binaries we generate. Before this
patch the tpm2_verifysigniture executable weighed in at 162392. After
this patch it was over 20k smaller (using standard flags on Debian).

Signed-off-by: Philip Tricca <flihp@twobit.us>