Add machinery to assist in parsing command line options and creating contexts (TCTIs & SAPI).
Overview:
The options & context-util modules allow us to support multiple TCTIs
dynamically at the time one of our tools is executed. Additionally we
include the build machinery necessary to enable / disable TCTIs at build
time for added flexibility. This current implementation preserves
existing default behavior of defaulting to the socket TCTI. The
down-side to this approach is that each tool must be ported over to use
these modules / patterns. This work will be followed by some new tools
that use these features, and an effort to port the existing tools.

src/options.c:
This module contains constants, structures and functions for obtaining
configuration information from the environment and the command line
(options on the commandline override the environment). This includes
simple stuff like the --help command line options as well as more
complex stuff like TCTI selection and configuration. The TCTI
libraries supported by this module are configurable at build time
through --with/--without options in the configure script.

src/context-util.c:
This module processes the common_opts_t structure from the options
module described above. It uses this data to instantiate and configure
the requested TCTI context. Additionally a SAPI context can be
instantiated with the requested TCTI. The supported TCTIs are
configurable through the configure script.

src/main.c
This is a simple command wrapper. It implements a 'main' function that
parses all of the common command line options, packages them up into a
structure, and instantiates the appropriate TCTI & SAPI context. It
also automates dumping the right manual for commands (that have
manuals) when the --help option is provided. A tool built on this
pattern will implement the 'execute_tool' function and build with the
src/main.c file to get an executable.

man/*.troff:
Manuals / "man" pages will need to include a description of all the
common and TCTI options. Duplicating this information across all
manuals would be a huge maintenance burden. We provide two troff
fragments that should be included in each manual using the rule in the
Makefile.am for construction. The method used to insert these fragments
is a bit of a hacked up version of '#include' but I was unable to get
troff '.so' style includes to work when the 'man' command is invoked.
All source for the manpages will be in man/ but when built, the output
will be in man/man8/ so that MANPATH=$(srcdir) can be used for testing.

Signed-off-by: Philip Tricca <flihp@twobit.us>