tpm2_send_command: Add tool to send a command provided on stdin to the TPM.
This command reads a TPM command buffer from its standard input stream
and then sends the command to the TPM. The command is transmitted using
the TCTI information provided by the caller on the command line. The
response received from the TPM is written to stdout. This command is
intended for debugging / testing usage in situations where the caller may
have a binary TPM command saved in a file on disk and they want to send
it to the TPM saving the results in another file. To do so would look
something like this:
tpm2_send_command [options] < tpm2-command.bin > tpm2-response.bin

We include a test to invoke this in the test/ directory as well. This
includes a data file continaing a simple TPM2 command buffer to get the
TPMs fixed properties using the GetCapability command.

Signed-off-by: Philip Tricca <flihp@twobit.us>