tpm2_startup: Add command to send TPM2_Startup with required options.
This command doesn't make a lot of sense for use with the integrated TPM
on most systems since your firmware should be doing this for you. But if
you're debugging against a TPM2 software simulator then every time you
restart it you'll need to execute a TPM_Startup and this tool should
make that a lot easier.

Provide either --clear or --state depending on which Startup command you
want to execute. All of the TCTI options should work as well so this can
be sent to the physical TPM if you need it. All of this is documented in
the provided man page too.

Signed-off-by: Philip Tricca <flihp@twobit.us>