context-util: Deallocate context memory after failed initialization calls.
This is a hazard of the model for TCTI and SAPI context initialization.
It's possible for the first call to succeed while the second call fails.
In this case we must free the allocated data before we return an error
indicator (a NULL pointer). Otherwise we leak the memory we allocated
for the context structure.

There are 3 cases where this may happen in the context-util module. One
was already handled properly. This commit fixes up the other two.

Signed-off-by: Philip Tricca <flihp@twobit.us>