options: Remove common options from argument vector after processing.
This patch fixes a few issues:
Firstly adding the '-' to the beginning of the options string prevents
getopt from moving what it thinks are positional parameters to the end
of the argument vector. This caused a bug where tool specifc options
would be rearranged preventing tools from functioning properly.

Secondly this patch adds logic to remove TCTI / common options from the
argument vector after they're parsed. This keeps the parsing logic in
each tool from having to be smart enough to ignore these additional
options.

This last bit is a bit more complex in the implementation than it
sounds since we must allocate a new argument vector and populate it with
the arguments / options that the common processing code skips over. This
vector must then be freed after the tool finishes executing. This
complexity is not visible to the tool code itself though.

Signed-off-by: Philip Tricca <flihp@twobit.us>