tpm2_create: Port to new tcti / options infrastructure.
This one was pretty straightforward after the refactoring done in
porting the tpm2_createprimary tool. All we do here is convert the main
function to execute_tool, remove the -h/-v options (and all related
logic), and move the rule to build this tool outside the HAVE_TCTI_SOCK
conditional in the Makefile.am.

Signed-off-by: Philip Tricca <flihp@twobit.us>