string-bytes: re-implement string to value routines
getSizeUintXX and getSizeUintXXHex routines were very similair.
Refactor them to use strtoul(), and allow strtoul to handle the base.
Also, call the 32 bit variant from the 16 and just check for overflow.

Mark the old interfaces as deprecated, since there are a TON of callers and
convert them on later porting/refactoring endeavors.

Signed-off-by: William Roberts <william.c.roberts@intel.com>