configure: use git describe for version number
Remove the hardcoded version number and use git describe. This makes
it impossible to release a version of the tools that does not agree
with the git tag.

Additionally, builds not on a tag will be marked dirty:

$ ./src/tpm2_listpcrs -v
./src/tpm2_listpcrs, version v1.1.0-81-ga5d2946

Once the current HEAD is a tag, it will only read the tag number.

Signed-off-by: William Roberts <william.c.roberts@intel.com>