split apart session to remove usages of hardcoded sys context
Split apart the session code so that the sapi context is not the
global, initialized one, but rather can be passed along from
tool_execute. This allows for tools to use the session API
without faling on the contetx initialization.

Introduce 3 new versions of old interfaces in:
tpm_hmac
tpm_session
tpm_kdfa

Mark the old ways as deprecated.

Signed-off-by: William Roberts <william.c.roberts@intel.com>