configure: add --enable-hardening support
Add CFLAGS/LDFLAGS by default to the build that are associated
with good security practices.

The configure scripts checks that all the flags are supported
and if something is not supported, prints an error message
indicating to configure with --disable-hardening.

If one configures with --disable-hardening an error message
indicating that it is dangerous and that they should fix the
configure script and post patches upstream is prevalent.

Signed-off-by: William Roberts <william.c.roberts@intel.com>