tpm2_getrandom: update help and simplify interface
Make argument options optional for tpm2_getrandom. Size is a required
argument, so make it a required positional arg. Now, the only option
is -o or --output, which optionally specifies an output file. If an
output file is specified, it writes all output to that file, else
it prints hex values of the bytes to the terminal.

Fixes 251

This patch requires a significant version bump as this version of
tpm2_getrandom is not "api" compatable with older versions.

Signed-off-by: William Roberts <william.c.roberts@intel.com>