tpm_tools: start adding more commands
Added:
akparse
activatecredential

This showed the conflict on having execute_tool defined in the source
file always. Thankfully, automake is smart enough to recompile sources
it the compilation context differs. We define a new ENTRY_POINT macro
so based on SHELL_TOOLS define, the compiation unit gets built for
a standalone tool or a shell tool.

This patch lays the foundation for just adding the rest of the
tools.

Signed-off-by: William Roberts <william.c.roberts@intel.com>