tpm2_getmanufec: stop strcpy into b64 buffer
The URL is being truncated when escaping causes the escaped
URL to be bigger than the b64 buffer.

curl_easy_escape() provides a heap allocated string that could
be larger than the buffer that the openssl bio routines use.

Rather than strncpy() it into that buffer, just return the
curl string.

Signed-off-by: William Roberts <william.c.roberts@intel.com>