spike for getrandom in memory return of value.
Currently needed output is going to stdout, which makes the scripting
via lua painful. I see no good way of grabbing stdout and reading it.

Rather than relying on stdout as the method for value return, have
things return a TPM return code along with a table. The table can
than be interogated for the value.

In normal tool mode, the output is still stdout, but now has an
easier to parse <key>:<value> syntax.

For example:
$ ./tpm2_getrandom 4
random:0x6E8C9558

The nice part about this, is output can now be parsed in a
meaningful way.

Signed-off-by: William Roberts <william.c.roberts@intel.com>