lib: fallback to other TCTIs if default TCTI can't be used
There is dynamic TCTI support to either choose at build or execution time
the TCTI to be used. But would be good to make this even more transparent
to the user by allowing the TCTI context creation to fallback to the other
TCTI if creating a context with the default fails.

For example if default TCTI is tabrmd but the tpm2-abrmd resource manager
hasn't been started, the tpm2.0-tools can at least work accessing the TPM
directly by using the device TCTI.

The fallback order is the same as used for deciding the default TCTI at
build time. If the default is tabrmd, then the socket and device TCTI are
tried and if the default is the socket TCTI then the device TCTI is tried.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>