tpm2_create: allow read data to be sealed from the standard input
The tool has an option to pass a file with data to be sealed. But in some
cases it may be important to never store the sensitive data and instead
allow tpm2_create to read the sensitive data from its standard input, e.g:

$ tpm2_create -g 0x000B -G 0x0008 -o pub.key -O priv.key -c primary.context -I- <<< $secret

Many command line utilities use "-" as a convetion to denote that the data
should be read from stdin instead of a file, so use the same convetion and
document it in the tpm2_create manual page.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>