test_tpm2_unseal: extend test to cover unsealing objects using a PCR policy
The test currently only tests creating and unsealing objects without a PCR
policy, extend the test to also cover the case when a PCR policy is used.

This also tests tpm2_create ability to get the data to be sealed from the
stdin and for the tpm2_unseal tool to write the unsealed data to stdout.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>