tpm2_unseal: use endorsement instead of platform hierarchy for test
The platform hierarchy is intended to be under the control of the platform
firmware, and so the test may not be usable in its current form on all the
systems. For example if there's a password for the platform hierarchy that
is set and the user has no way to clear it.

Change the test to use the endorsement hierarchy instead so the test works
on most systems.

Fixes #255

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>