coverity: skip building and testing
When coverity starts a scan, it builds first, thus the script
for jenkins CI fails becuase the top level tpm2.0-tools folder
has had --configure executed. Just skip building and testing
when pushing to coverity_scan, as we just want coverity to
build it.

We move the build commands to a script. Which has the nice
property of being able to run it locally:
$ .ci/travis-build-and-run-tests.sh

The downside is, individual command errors are not reported nicely
like the Travis CI script did.

Also, add ./bootstrap to coverity addon, as this in no longer in the
before_script tag. Coverity runs inbetween before_script and script.

The other option would be to:
1. Detect coverity scan branch and call make distclean from top level
2. Modify coverity to use the same build directory

Signed-off-by: William Roberts <william.c.roberts@intel.com>