main: add tpm2_startup() command to all tools
A common issue when using a device directly, is that, potentially,
no startup command was issued. A tpm cannot be used until a
startup command is issued.

The tools do not call startup implicitly, so add that to the tools
and ignore initialized or session errors. However, log at the
warning level any other return codes.

Fixes: #296

Signed-off-by: William Roberts <william.c.roberts@intel.com>