tpm2_createpolicy: don't end auth session for extend_policy_session
The build_policy() function has a boolean extend_policy_session parameter
that's used to decide if the policy session is an "extended" one. This
means that the function won't flush the policy session on exit so it can
be used by tools after bulding a policy.

But the function removes the session unconditionally, so also only remove
the session when the caller passed a false extend_policy_session argument.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>