tpm2_createpolicy: move out createpolicy ad-hoc logic from build_policy()
The build_policy() policy function logic is quite generic, but it has some
custom logic that's only related to the tpm2_createpolicy that prevents it
to be reused by other tools. For example it prints the policy digest when
using TPM_SE_POLICY and saves it to a file when using TPM_SE_TRIAL.

Move this custom logic out of build_policy() so the function can be reused.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>