tpm2_createpolicy: remove redundant .is_raw_pcrs_file field
The struct tpm2_pcr_policy_options has a boolean .is_raw_pcrs_file field
that's used to inform if a raw_pcrs_file was passed as an option.

But this is not necessary since is enough to check if the raw_pcrs_file
char pointer is not NULL. So just remove the redundant field.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>