tpm2_createpolicy: check earlier if PCR bank ids were selected
When building a PCR policy, a check is made if the list of PCR bank ids
were set using the -L option. The check can happen much earlier so move
it outside the build_pcr_policy() function. That way this function does
not depend on this flag that's specific to the tpm2_createpolicy tool.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>