tpm2_create: check if keyed-hash algorithm was used when sealing data
The TCG specification is clear that only the TPM_ALG_KEYEDHASH algorithm
can be used when the user provides part of the the sensitive data.

The tpm2_create tool behaves according to the specification, but it may
not be clear to the users why it fails, unless are already familiar with
the specs. Make the tool gracefully fail and also add a note to man page.

Fixes: #369

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>