tpm2_getmanufec: handle output more gracefully
Currently, the -E option will fail when -E is not specified:
$ tpm2_getmanufec -g 0x01 -O -N -U -f test_ek.pub -S https://ekop.intel.com/ekcertservice/
...
ERROR: Could not open file for writing: "(null)"

Make -E option, optional by making the default stdout. This requires
modifying the tool to stop outputing other information to sdout,
which is currently unused output in the test suite.

Make the information only output when verbose is enabled, this
includes curl verbose information.

Update errors and warnings to go to stderr.

Fixes: 361

Signed-off-by: William Roberts <william.c.roberts@intel.com>