tpm2_create: simplify execute_tool() error codes propagation
The tool main function that calls execute_tool() doesn't use the returned
value, it only takes 0 as a succeeds value an everything else as an error.

So it makes no sense to have a gazillion different error codes for each
situation, this just makes the code difficult to read for no gain.

While being there, handle the cases for when getopt_long() finds that the
option or arguments are not correct.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>