tpm2_getmanufec: simplify execute_tool() error codes propagation
The tool main function that calls execute_tool() doesn't use the returned
value, it only takes 0 as a succeeds value an everything else as an error.

So lets just return consistently 1 as an error an 0 as a success, instead
of returning a combination of -1 and value set in the return_val variable.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>