tpm2_nvlist: fix tool exit code
The tpm2_nvlist execute_tool() is wrongly return nv_list() != 0. This is
wrong not only because that function returns a bool and not an int, but
also because the nv_list() returns true (!= 0) on success and the main
the main program takes 1 to be an error. So fix both issues.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>