ci: Enable ASAN when building with clang
Use CLANGs asan feature to find memory issues during the CI
process.

Clang ASAN info can be found at:
https://clang.llvm.org/docs/AddressSanitizer.html

Also, disable test:
test_tpm2_getmanufec:

When using ASAN. ASAN is reporting a leak in tpm2_getmanufec
that cannot be suppressed with version 3.6 of clang. It is
undiagnosed if it is a bug in tpm2_getmanufec, OpenSSL or
a false positive.

For now, skip the test, as checking everything else is more
prident, and it gets functional tested under GCC builds.

File a bug to come back and try to fix it better:
https://github.com/01org/tpm2-tools/issues/390

Fixes: #387

Signed-off-by: William Roberts <william.c.roberts@intel.com>