lib: rename file_load_tpm_context_from_file() to follow coding standard
The coding standard document says that a function that is exported should
be prefixed with the source file (or object module) name it was defined.

But file_load_tpm_context_from_file() is using file as prefix instead of
files. This is trivial but since there is a bug filled for this, fix it.

Fixes: #367

Reported-by: 刘群 <qunliu@zyhx-group.com>
Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>