coveralls: fix uploads
Coveralls was not being updated as ENABLE_COVERAGE was exported in
a seperate environment and not preserved when invoking the coveralls
upload script. Move the coveralls upload invocation within that
environment.

Also, the COVERALLS_REPO_TOKEN is not set by travis when building
PRs, so make it a non fatal error if it's not set

Signed-off-by: William Roberts <william.c.roberts@intel.com>