version: display configured tcti
Having only the manpage contain the mapping of configured tctis can be quite
challenging, if one is trying to figure out how this software is confogured
in an environment where manpages are not available. Add, to the version
a list of configured tctis, so a simple -v provides supported, configure
information.

This information will help bug reports, as a version of the software with
different configurations may have different bugs.

Signed-off-by: William Roberts <william.c.roberts@intel.com>