tpm2_pcrevent: add tool
Add the tpm2_pcrevent tool. It uses either the
TPM2_PCREvent if the data to hash is under 1024
and uses the sequence API if it is larger.

This command can optionally take a pcr index and perform
a pcrextend on the digest data.

The man page is a departure from the raw troff manpages
and now uses md2man (mark down to man).

Add a test script for testing tpm2_pcrevent.

Signed-off-by: William Roberts <william.c.roberts@intel.com>