test: fix -I- option to tpm2_create
The fact that the option parsing worked with -I- was a sideaffect of
always reallocing the new argv array, and no proper size check.

During this test, the argv from the OS contained 17 entries.
After handling the common options, it was 18, with a duplicate
of the optarg from the option before -i-.

There needs to be a space beteen a single option letter and its
value.

Signed-off-by: William Roberts <william.c.roberts@intel.com>