test: fix shebang interpreter magic
The kernel checks the first two bytes of a file for certain
magic values. If they are a particular value, the kernel
invokes the path after the magic (ie the interpreter).

The shell scripts started with a comment block, so when
passed directly to an exec family of calls, they fail.

Fix this by putting the shebang magic at the begining,
before the license stanza.

Signed-off-by: William Roberts <william.c.roberts@intel.com>