tpm2_hash: Drop -I option in man page example
The commit 6388831 ("tpm2_hash: drop -I option") changed tpm2_hash tool
to get the data to hash from either the standard input or a file. So it
removed the -I option.

But forgot to update the example in the man page. While being there, do
as other examples in man pages and show a specific command instead of a
generic one.

Fixes: #441

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>