test: Use relative path for yaml_get.py in test_tpm2_pcrevent.sh
The test_all.sh script adds the tests directory in the $PATH variable but
that's not the case when the tests are executed standalone. In this case,
the test_tpm2_pcrevent.sh doesn't find the yaml_get.py and fails with:

$ ./test_tpm2_pcrevent.sh
./test_tpm2_pcrevent.sh: line 74: yaml_get.py: command not found
old_pcr_value=`yaml_get.py sha1 9 $yaml_out_file` on line 74 failed: 127

Tests are expected to be executed inside the test/system sub-directory,
so just use a relative path for the yaml_get.py script.

While being there, add a printout to notify that the test was executed
successfully to be consistent with the other tests that do the same.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>