man: explain in tpm2_evictcontrol page how objects are evicted
The specification says "If objectHandle is a persistent object, then the
call evicts the persistent object" and "if objectHandle is a persistent
object handle, then it shall be the same value as persistentHandle".

So the tool is doing the correct thing by requiring both the object and
persistent handles to be defined and being the same to evict an object.

But this isn't documented in the tool man page and it may not be evident
to people not familiar with the TCG TPM specification.

While being there, remove a -X option from one of the examples since the
tpm2_evictcontrol tool doesn't have one.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>