tpm2_evictcontrol: make tool smarter to not require -S to evict objects
The TCG TPM specification requires the same handle to be provided for both
objectHandle and persistentHandle but this implementation detail leaked in
the tool and so users need to provide both a -S and -H options to evict an
object.

But the tool can be smarter, the handles are not just opaque values but
they encode information. The first byte is the TPM_HT (Handle Type) and
for persistent objects handles this is TPM_HT_PERSISTENT (0x81).

Fixes: #492

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>