tpm2_load: don't make -n option to be required
The tool requires the option for a file path to store the structure name
of the object, but this shouldn't really be a required option. In fact,
the whole option checking seems bogus so just check for the options that
should be required and not the ones that should be optional.

While being there, make this clear in the man page as well.

Fixes: #494

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>