tpm2_listpersistent: check that -g and -G correct values were provided
The tool is using the tpm2_alg_util_from_optarg() function to convert from
the pretty names provided by the user to the hexadecimal algorithm values.

But the same function is used to convert both key and hash algorithms and
so using a key as a hash and viceversa was wrongly taken as correct value.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>