tpm2_getmanufec: fix NULL pointer dereference due missing password options
The tool has options to specify passwords for the endorsement hierarchy,
endorsement primary key and the owner hierarchy. The first two are used
if the EK has to be created (when is not offline retrieved) and the last
one is used if the EK has to be made persistent in the TPM2.

All these passwords options are optional, but the tool assumes that are
always set and so attempts to access the NULL pointers:

$ tpm2_getmanufec
Segmentation fault (core dumped)

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>