tpm2_getmanufec: remove new line characters in log message strings
Many log messages in the tool start with a new line character, but this
is not needed since the LOG_* macros ensures that a new line will always
be printed at the end of a log message.

In fact, it's wrong to do it since the macro prefix the log level to the
message, so starting with a '\n' will print messages like the following:

$ tpm2_getmanufec
ERROR:
EvictControl:Make EK persistent Error. TPM Error:0x1c4

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>