man: use pandoc instead of md2man for man pages generation
The tools man pages used to be static roff files, and were converted to be
generated by md2man-roff from markdown documents.

But md2man is a rather obscure project, no major distros have the ruby gem
packaged, is basically a one man project and its last release was almost 2
years ago. Finally, it depends on the binman and opener projects from the
same author that have even older releases and less development.

While packaging these ruby gems is trivial, the fact that are no packages
is an indication that no many projects use it.

The pandoc project on the other hand has hundreds of contributors, it's in
active development and all major Linux distributions have packages for it.

So let's use pandoc to generate the man pages instead of md2man. This will
make easier not only for distros to update the tpm2-tools package but also
for developers and users to build tpm2-tools (with man pages) from source.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>