test: tpm2_createpolicy: explicitly define a base when comparing hex values
If the base is omitted in an arithmetic expression bash assumes a base 10,
unless the constant starts with a 0 or 0x. In these cases, base 8 and 16
are used respectively.

So the test was only working because assumes a PCR0 value of 3, which
has the same value in both bases, but this isn't true for all values and
will lead to the following issue:

./test_tpm2_createpolicy.sh: line 57: [[:
9872964b9b40cdd0363fcd6af8c267c9cb34200b: value too great for base
(error token is "9872964b9b40cdd0363fcd6af8c267c9cb34200b")
Failure: Creating Policy Digest with PCR policy for index 0 and sha1
pcr index hash

Fix this by explicitly define a base 16 so bash know that the constant
value is an hexadecimal number.

Signed-off-by: Jia Zhang <qianyue.zj@alibaba-inc.com>