test: query the list of supported hash algorithms before calling tpm2_create
The unsupported hash algorithms should be ignored before the testing.
It is necessary to implement directly querying the supported algorithms
in form of either hex value of friendly name.

So let's have a common script test_helpers.sh and the tests that need it
can just source it.

This change partially resolves the issue #335 (Make sure that all tests
work on a real TPM hardware) for helping on it regard.

Signed-off-by: Jia Zhang <qianyue.zj@alibaba-inc.com>