tpm2_import: remove unneeded ifdefery to call the RSA_set0_key function
The ssl_RSA_set0_key() function already has an error path, but the error
was not propagated to the caller. Since the function has basically the
same signature than the OpenSSL 1.1.0 RSA_set0_key() function, define it
for OpenSSL < 1.1.0 so there is no need for the ifdefery in the caller.

Suggested-by: Matthias Gerstner <matthias.gerstner@suse.de>
Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>