test_tpm2_import: work around spec errata
TPM 2.0 spec 1.16 errata on decrypt and sign set for a symcipher.

Certain TPM 2.0 chip, e.g, Nationz Z32H320TC, and tpm2simulator
referring to TPM 2.0 spec 1.16 may have an errata, disallowing both
decrypt and sign set for a symcipher object, and in response to
RC_ATTRIBUTES.

In order to work around it, we attempt to run tpm2_importe again with
sign bit clear.

This change partially resolves the issue #335 (Make sure that all
tests work on a real TPM hardware) for helping on it regard.

Signed-off-by: Jia Zhang <qianyue.zj@alibaba-inc.com>