tpm2_errata: implement a generic handling for errata
The TPM errata should be addressed and handled transparently by design.

In order to handle an errata, the errata info should be recorded in
tpm2_errata.[ch], and the tools need to query the errata handler with
a pre-defined errata index number and fixup handler.

During the initialization, the TPM spec and errata info will be
remembered and used to check the errata query issued by the callers.

If the errata is omitted, nothing would happen. In contrast, the log
records which errata is applied.

Signed-off-by: Jia Zhang <qianyue.zj@alibaba-inc.com>