tpm2_unseal: make -F optional to allow using current PCR values for auth
The tool currently checks if both the -L and -F options were defined and
only builds a PCR policy session for authentication if that's the case.

But the tpm2_unseal man page says that the -F option is not required, if
this isn't defined then the current PCR values are used to build the PCR
policy used for authorization to unseal the data.

So honor the man page says and don't make the -F option to be required,
instead just always pass raw_pcrs_file to tpm2_policy_build(). If it's
NULL, then this function reads the current PCR values instead of trying
to use the values that are defined in a file.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>